+++ src/main/java/org/bytedeco/javacpp/BytePointer.java
@@ -77,7 +77,7 @@ public class BytePointer extends Pointer {
      */
     public BytePointer(ByteBuffer buffer) {
         super(buffer);
-        if (buffer != null && buffer.hasArray()) {
+        if (buffer != null && !buffer.isDirect() && buffer.hasArray()) {
             byte[] array = buffer.array();
             allocateArray(array.length - buffer.arrayOffset());
             put(array, buffer.arrayOffset(), array.length - buffer.arrayOffset());