+++ src/main/java/org/bytedeco/javacpp/CharPointer.java
@@ -60,7 +60,7 @@ public class CharPointer extends Pointer {
      */
     public CharPointer(CharBuffer buffer) {
         super(buffer);
-        if (buffer != null && buffer.hasArray()) {
+        if (buffer != null && !buffer.isDirect() && buffer.hasArray()) {
             char[] array = buffer.array();
             allocateArray(array.length - buffer.arrayOffset());
             put(array, buffer.arrayOffset(), array.length - buffer.arrayOffset());